+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [UNRELEASED]
+### Changed
+- Use v0.4.2 of `hydrogen/module.core`
 
 ## [0.5.10] - 2021-06-24
 ### Fixed